requirements.txt
Django
psycopg2
djangorestframework
djangorestframework-jwt
django_extensions